Add icu_compare() and icu_sort_key() variants using collation of the inputs.
These variant forms get the ICU collation spec from the pg_locale_t object
corresponding to the collation of the function call,
rather than as an explicit argument.

It's a simpler interface when the collation has been created in postgres.

And it's really faster due to Postgres caching the return of ucol_open()
for the duration of the session. icu_ext could cache it too but
on the whole it looks cleaner to not leak memory if we have a reasonable
alternative.